flask_main/requirements.txt
Flask
cx_Oracle
python-dotenv
requests
openai